+++ templates/BookProject/blocks/hidden.ps1
@@ -1,0 +1,3 @@
+﻿Register-EBMarkdownBlock -Name hidden -Converter {
+	param ($Data)
+}